V1 Notes

+++ src/pages/HomePage.jsx
@@ -5,6 +5,7 @@ import NoteList from "../components/NoteList";
 import SearchBar from "../components/SearchBar";
 
 import { deleteNote, getActiveNotes, archiveNote } from "../utils/local-data";
+import AddButton from "../components/AddButton";
 
 function HomePageWrapper() {
   const [searchParams, setSearchParams] = useSearchParams();
@@ -21,6 +22,11 @@ function HomePageWrapper() {
 }
 
 class HomePage extends React.Component {
+  static propTypes = {
+    defaultKeyword: PropTypes.string,
+    keywordChange: PropTypes.func.isRequired,
+  };
+
   constructor(props) {
     super(props);
 
@@ -35,14 +41,20 @@ class HomePage extends React.Component {
   }
 
   onDeleteHandler(id) {
-    deleteNote(id);
+    const isConfirmed = window.confirm(
+      "Apakah yakin untuk menghapus notes ini?"
+    );
 
-    // update the note state from data.js
-    this.setState(() => {
-      return {
-        notes: getActiveNotes(),
-      };
-    });
+    if (isConfirmed) {
+      deleteNote(id);
+
+      // update the note state from data.js
+      this.setState(() => {
+        return {
+          notes: getActiveNotes(),
+        };
+      });
+    }
   }
 
   onKeywordChangeHandler(keyword) {
@@ -78,8 +90,8 @@ class HomePage extends React.Component {
       });
 
     return (
-      <section>
-        <h2>Daftar Notes</h2>
+      <section className="homepage">
+        <h2>Daftar Catatan</h2>
         <SearchBar
           keyword={this.state.keyword}
           keywordChange={this.onKeywordChangeHandler}
@@ -89,6 +101,9 @@ class HomePage extends React.Component {
           onDelete={this.onDeleteHandler}
           onArchive={this.onArchiveHandler}
         />
+        <div className="homepage__action">
+          <AddButton />
+        </div>
       </section>
     );
   }
